+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 DevOps Tool: Core
 ==============================================
 
-# 0.1.1 (Unreleased)
+# 0.2.0 (Unreleased)
 - Added DatabaseAdapterManager
 - Merged DatabaseMetadataProviderInterface into DatabaseAdapterInterface
 - Added exception handling around entire app
 - Added FilesystemAdapterManager
 - Added filesystem:ls, filesystem:copy, and filesystem:sync commands
+- Added MountManager with SyncPlugin
+- Added amphp/amp for parallel processing
+- Applied parallel processing to SyncPlugin
 
 # 0.1.0
 - Initial build copied over from DevOps tool